requirements.txt
# Core ML & Data Science
torch==2.2.2
transformers==4.35.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3

# Vector Database
faiss-cpu==1.7.4

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
pydantic==2.5.0

# Utilities
tqdm==4.66.1
requests==2.31.0
python-multipart==0.0.6
pyarrow==14.0.1

# Development
jupyter==1.0.0
ipykernel==6.26.0